Moved types to a shared place
