+++ package.json
@@ -13,7 +13,7 @@
   "dependencies": {},
   "devDependencies": {
     "babel-core": "^6.23.1",
-    "babel-jest": "^20.0.0",
+    "babel-jest": "^20.0.3",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.22.0",
