+++ packages/common/src/types.js
@@ -13,3 +13,8 @@ export type QueryData = {
   conditionFunction: Function,
   elements: Array<ElementData>
 };
+
+export type ContainerSize = {
+  width: number,
+  height: number
+};
