+++ packages/container-query/.babelrc
@@ -1,5 +1,5 @@
 {
     "presets": [
         ["es2015", { "modules": false }]
-    ],
+    ]
 }
