+++ packages/container-query/.flowconfig
@@ -1,6 +1,8 @@
 [ignore]
+./dist
 
 [include]
+../common/src
 
 [libs]
